# Adjust histogram buckets for faster responses