--- a/handlers/metrics.go
+++ b/handlers/metrics.go
@@ -33,7 +33,7 @@ var (
 			Name: "router_backend_handler_response_duration_seconds",
 			Help: "Histogram of response durations by router backend handlers",
 			Buckets: prometheus.ExponentialBuckets(
-				0.25, 2, 6, // This buckets [...0.25  0.5  1  2  4  8...]
+				0.0625, 2, 7, // This buckets [...0.0625  0.125  0.25  0.5  1  2  4...]
 			),
 		},
 		[]string{